refactor(fee-dialog): move and rename FeeLimitDialog to FeeConfigDialog
