+++ src/components/settings/SettingsPage.tsx
@@ -35,10 +35,10 @@ import { authStore } from '@/store/authStore'
 import { jamSettingsStore } from '@/store/jamSettingsStore'
 import { Spinner } from '../ui/spinner'
 import { AccountXpubsDialog } from './AccountXpubsDialog'
-import { FeeLimitDialog } from './FeeLimitDialog'
 import { LanguageSelector } from './LanguageSelector'
 import { SeedPhraseDialog } from './SeedPhraseDialog'
 import { SettingItem, SettingsLink, SettingSwitch } from './SettingsItem'
+import { FeeConfigDialog } from './fees/FeeConfigDialog'
 
 interface SettingPageProps {
   walletFileName: WalletFileName
@@ -55,7 +55,7 @@ export const SettingsPage = ({ walletFileName, onLockWallet }: SettingPageProps)
 
   const [showSeedDialog, setShowSeedDialog] = useState(false)
   const [showXpubsDialog, setShowXpubsDialog] = useState(false)
-  const [showFeeLimitDialog, setShowFeeLimitDialog] = useState(false)
+  const [showFeeConfigDialog, setShowFeeConfigDialog] = useState(false)
   const hashedPassword = useStore(authStore, (state) => state.state?.hashed_password)
   const { isFeatureEnabled } = useFeatures()
 
@@ -125,7 +125,7 @@ export const SettingsPage = ({ walletFileName, onLockWallet }: SettingPageProps)
           <SettingItem
             icon={DollarSignIcon}
             title={t('settings.show_fee_config')}
-            action={() => setShowFeeLimitDialog(true)}
+            action={() => setShowFeeConfigDialog(true)}
           />
         </CardContent>
       </Card>
@@ -248,11 +248,11 @@ export const SettingsPage = ({ walletFileName, onLockWallet }: SettingPageProps)
         </>
       )}
 
-      <FeeLimitDialog
-        key={`fee-dialog-open-${showFeeLimitDialog}`}
+      <FeeConfigDialog
+        key={`fee-dialog-open-${showFeeConfigDialog}`}
         walletFileName={walletFileName}
-        open={showFeeLimitDialog}
-        onOpenChange={setShowFeeLimitDialog}
+        open={showFeeConfigDialog}
+        onOpenChange={setShowFeeConfigDialog}
       />
 
       {hashedPassword && (
