+++ src/components/settings/fees/CollaboratorFeesForm.tsx
@@ -3,9 +3,9 @@ import { type UseFormReturn } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 import { Alert, AlertDescription } from '@/components/ui/alert'
 import { cn } from '@/lib/utils'
-import { Field, FieldDescription, FieldLabel } from '../ui/field'
-import { InputGroup, InputGroupAddon, InputGroupInput } from '../ui/input-group'
-import { SatSymbol } from '../ui/jam/CurrencySymbol'
+import { Field, FieldDescription, FieldLabel } from '../../ui/field'
+import { InputGroup, InputGroupAddon, InputGroupInput } from '../../ui/input-group'
+import { SatSymbol } from '../../ui/jam/CurrencySymbol'
 import type { CollaboratorFeesFormValues } from './CollaboratorFeesForm.schema'
 
 const FieldPrefixSatSymbol = (
